backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
werkzeug==3.0.1
python-magic==0.4.27
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
alembic==1.13.1
python-dotenv==1.0.0
gunicorn==21.2.0
pymysql==1.1.0
cryptography==41.0.8